Proof of concept trajectory class iterators
This is a proof of concept that replaces the iterators returned by the
proto reader __getitem__ by an iterable class. That class has a __len__
method to adress #1894 and can have a write method to adress #1037.

In the current implementation, each type of input for the proto reader's
__getitem__ returns a different class. This allow to test for the
correct behavior only once. Ideally, the iterables should be imutable to
avoid inconsistencies.